Update query counting test
